refactor: 활동보고서 상세조회 컴포넌트 분리
